Spring Data

+++ AdvertisementProject/service/src/main/java/com/senlainc/javacourses/petushokvaliantsin/service/payment/PaymentTypeService.java
@@ -1,11 +1,13 @@
 package com.senlainc.javacourses.petushokvaliantsin.service.payment;
 
 import com.senlainc.javacourses.petushokvaliantsin.dto.payment.PaymentTypeDto;
+import com.senlainc.javacourses.petushokvaliantsin.enumeration.EnumException;
 import com.senlainc.javacourses.petushokvaliantsin.enumeration.EnumLogger;
 import com.senlainc.javacourses.petushokvaliantsin.model.payment.PaymentType;
 import com.senlainc.javacourses.petushokvaliantsin.repositoryapi.payment.IPaymentTypeDao;
 import com.senlainc.javacourses.petushokvaliantsin.service.AbstractService;
 import com.senlainc.javacourses.petushokvaliantsin.serviceapi.payment.IPaymentTypeService;
+import com.senlainc.javacourses.petushokvaliantsin.utility.exception.EntityNotExistException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,6 +41,9 @@ public class PaymentTypeService extends AbstractService implements IPaymentTypeS
     @Override
     @Transactional
     public boolean delete(Long index) {
+        if (!paymentTypeDao.existsById(index)) {
+            throw new EntityNotExistException(String.format(EnumException.ENTITY_NOT_EXIST.getMessage(), "Type", "id", index));
+        }
         paymentTypeDao.deleteById(index);
         LOGGER.info(EnumLogger.SUCCESSFUL_DELETE.getText());
         return true;
@@ -47,8 +52,10 @@ public class PaymentTypeService extends AbstractService implements IPaymentTypeS
     @Override
     @Transactional
     public boolean update(PaymentTypeDto object) {
+        if (!paymentTypeDao.existsById(object.getId())) {
+            throw new EntityNotExistException(String.format(EnumException.ENTITY_NOT_EXIST.getMessage(), "Type", "id", object.getId()));
+        }
         final PaymentType paymentType = dtoMapper.map(object, PaymentType.class);
-        paymentTypeDao.findById(paymentType.getId());
         paymentTypeDao.save(paymentType);
         LOGGER.info(EnumLogger.SUCCESSFUL_UPDATE.getText());
         return true;
